Rename InputCreatePassword to InputTakeNewPassword
